\name{adehabitatMA-internal}
\alias{showspidf}
\alias{showspodf}
\alias{showsp}
\alias{showsgdf}
\alias{head.SpatialPoints}
\alias{head.SpatialPointsDataFrame}
\alias{tail.SpatialPoints}
\alias{tail.SpatialPointsDataFrame}
\title{Internal adehabitatMA objects}
\description{Internal adehabitatMA objects.}
\details{These are not to be called by the user.}
\keyword{internal}