\name{show-methods}
\docType{methods}
\alias{show,SpatialGridDataFrame-method}
\alias{show,SpatialPixels-method}
\alias{show,SpatialPixelsDataFrame-method}
\alias{show,SpatialPolygonsDataFrame-method}
\title{ Print Functions for Objects of the Package 'sp' Used in 'adehabitatMA'}
\description{
  These are functions allowing a shorter print of the content of objects
  of class \code{SpatialPolygonsDataFrame} ,
  \code{SpatialPixelsDataFrame}, \code{SpatialPixels}, and
  \code{SpatialGridDataFrame}.  Original printing methods can be
  recovered by setting \code{adeoptions(shortprint=FALSE)}.
}
\section{Methods}{
\describe{
\item{object = "SpatialGridDataFrame"}{show function for the class \code{SpatialGridDataFrame}.}

\item{object = "SpatialPixels"}{show function for the class \code{SpatialPixels}. }

\item{object = "SpatialPixelsDataFrame"}{show function for the class \code{SpatialPixelsDataFrame}.}

\item{object = "SpatialPolygonsDataFrame"}{show function for the class \code{SpatialPolygonsDataFrame}.}

}}
\keyword{methods}
